Some work on packets compression
- Use libzstd for nodedef, itemdef, nodemeta if possible
- Compress inventory, inventory formspec, request media and media
  announcement packets
- Use compression for chat
- Compress also active object packets
- Try also with single player for now

For the chat, for now compress when sender length + message length is greater
than 64 (wstring). Compressed messages have higher version number so the
client knows if it should decompress. For TOSERVER message version number
has been added.

For the active object packets, it sometimes spams a lot when player is near for
example animal mobs. And it can be compressed with good ratio.

It seems it doesn't have much sense to compress also other packets because
they are either very small (so very little or no gain) or they are sent rarely.